requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==4.4.2
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
imageio==2.19.0
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.22.3
Pillow==9.1.0
proglog==0.1.10
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tqdm==4.64.0
urllib3==1.26.9
